# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RERA Agents API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Agent {
  id: number;
  /** @nullable */
  serialNumber?: number | null;
  name: string;
  agentType: string;
  /** @nullable */
  designation?: string | null;
  /** @nullable */
  personName?: string | null;
  /** @nullable */
  registeredAddress?: string | null;
  /** @nullable */
  phoneNumber?: string | null;
  /** @nullable */
  emailId?: string | null;
  /** @nullable */
  registrationNo?: string | null;
  /** @nullable */
  validUntil?: string | null;
  /** @nullable */
  certificateUrl?: string | null;
  /** @nullable */
  renewalCount?: number | null;
  scrapedAt: string;
  createdAt: string;
  updatedAt: string;
}

export interface AgentListResponse {
  agents: Agent[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface AgentStats {
  total: number;
  individuals: number;
  companies: number;
  expired: number;
  active: number;
  /** @nullable */
  lastScrapedAt?: string | null;
}

export interface ScraperResult {
  success: boolean;
  agentsScraped: number;
  agentsInserted: number;
  agentsUpdated: number;
  message: string;
  /** @nullable */
  error?: string | null;
}

export interface ScraperStatus {
  /** @nullable */
  lastRunAt?: string | null;
  totalAgents: number;
  isRunning: boolean;
}

export interface RecaMember {
  id: number;
  /** @nullable */
  photoUrl?: string | null;
  name: string;
  /** @nullable */
  age?: string | null;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  membershipId?: string | null;
  /** @nullable */
  reraNo?: string | null;
  /** @nullable */
  hiraNo?: string | null;
  /** @nullable */
  gstNo?: string | null;
  /** @nullable */
  mobileNo?: string | null;
  /** @nullable */
  altMobileNo?: string | null;
  /** @nullable */
  landline?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  associationName?: string | null;
  /** @nullable */
  coreCompetence1?: string | null;
  /** @nullable */
  coreCompetence2?: string | null;
  /** @nullable */
  coreCompetence3?: string | null;
  /** @nullable */
  coreCompetence4?: string | null;
  scrapedAt: string;
  createdAt: string;
  updatedAt: string;
}

export interface RecaMemberListResponse {
  members: RecaMember[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface RecaScrapeResult {
  success: boolean;
  scraped: number;
  inserted: number;
  updated: number;
  message: string;
  /** @nullable */
  error?: string | null;
}

export interface RecaScrapeStatus {
  isRunning: boolean;
  /** @nullable */
  lastRunAt?: string | null;
  totalMembers: number;
}

export type ListAgentsParams = {
  search?: string;
  agentType?: string;
  page?: number;
  limit?: number;
};

export type TriggerScrapeParams = {
  maxPages?: number;
};

export type ListRecaMembersParams = {
  search?: string;
  page?: number;
  limit?: number;
};