# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RERA Agents API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reca-members:
    get:
      operationId: listRecaMembers
      tags: [reca-members]
      summary: List RECA Kolkata individual members
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecaMemberListResponse"
...
  /reca-members/{id}:
    get:
      operationId: getRecaMember
      tags: [reca-members]
      summary: Get a single RECA Kolkata member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecaMember"
        "404":
          description: Not found
components:
  schemas:
...
    RecaMember:
      type: object
      properties:
        id:
          type: integer
        photoUrl:
          type: ["string", "null"]
        name:
          type: string
        age:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        membershipId:
          type: ["string", "null"]
        reraNo:
          type: ["string", "null"]
        hiraNo:
          type: ["string", "null"]
        gstNo:
          type: ["string", "null"]
        mobileNo:
          type: ["string", "null"]
        altMobileNo:
          type: ["string", "null"]
        landline:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        associationName:
          type: ["string", "null"]
        coreCompetence1:
          type: ["string", "null"]
        coreCompetence2:
          type: ["string", "null"]
        coreCompetence3:
          type: ["string", "null"]
        coreCompetence4:
          type: ["string", "null"]
        scrapedAt:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - scrapedAt
        - createdAt
        - updatedAt
    RecaMemberListResponse:
      type: object
      properties:
        members:
          type: array
          items:
            $ref: "#/components/schemas/RecaMember"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - members
        - total
        - page
        - limit
        - totalPages
...