# lib/api-zod/src/generated/types/recaMember.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RERA Agents API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RecaMember {
  id: number;
  /** @nullable */
  photoUrl?: string | null;
  name: string;
  /** @nullable */
  age?: string | null;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  membershipId?: string | null;
  /** @nullable */
  reraNo?: string | null;
  /** @nullable */
  hiraNo?: string | null;
  /** @nullable */
  gstNo?: string | null;
  /** @nullable */
  mobileNo?: string | null;
  /** @nullable */
  altMobileNo?: string | null;
  /** @nullable */
  landline?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  associationName?: string | null;
  /** @nullable */
  coreCompetence1?: string | null;
  /** @nullable */
  coreCompetence2?: string | null;
  /** @nullable */
  coreCompetence3?: string | null;
  /** @nullable */
  coreCompetence4?: string | null;
  scrapedAt: string;
  createdAt: string;
  updatedAt: string;
}