# lib/api-zod/src/generated/types/recaMemberListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RERA Agents API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RecaMember } from "./recaMember";

export interface RecaMemberListResponse {
  members: RecaMember[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}